requirements.txt
arabic-reshaper==2.1.1
Django==3.1.5
django-datatables-view==1.19.1
django-jet==1.0.8
django-phone-field==1.8.1
django-rest-swagger==2.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docopt==0.6.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
social-auth-app-django
social-auth-core
xhtml2pdf==0.2.5
pycountry
emoji-country-flag
googletrans==3.1.0a0
django_cron
django-filter
drf-flex-fields
channels
channels_redis
django-channels-presence